Re-structure backend server
