+++ server/business.js
@@ -1,0 +1,141 @@
+/**
+ * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import axios from "axios";
+import { getAccessToken, getOrganizationToken } from "./auth.js";
+import { agent, ASGARDEO_BASE_URL } from "./config.js";
+
+export async function isBusinessNameAvailable(businessName) {
+
+  const token = await getAccessToken();
+  const response = await axios.post(
+    `${ASGARDEO_BASE_URL}/api/server/v1/organizations/check-name`,
+    { name: businessName },
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+        Accept: "application/json",
+      },
+      httpsAgent: agent,
+    }
+  );
+  return response.data.available;
+}
+
+export async function createOrganization(businessName, creatorId, creatorUsername) {
+
+  const token = await getAccessToken();
+  const response = await axios.post(
+    `${ASGARDEO_BASE_URL}/api/server/v1/organizations`,
+    {
+      name: businessName,
+      attributes: [
+        { key: "creator.id", value: creatorId },
+        { key: "creator.username", value: creatorUsername },
+      ],
+    },
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+        Accept: "application/json",
+      },
+      httpsAgent: agent,
+    }
+  );
+  return response;
+}
+
+export async function getUserIdInOrganization(organizationId, username) {
+
+  const token = await getOrganizationToken(organizationId);
+  const response = await axios.get(
+    `${ASGARDEO_BASE_URL}/o/scim2/Users`,
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+        Accept: "application/json",
+      },
+      params: {
+          filter: `userName eq ${username}`,
+        },
+      httpsAgent: agent,
+    }
+  );
+
+  const resources = response.data.Resources || [];
+  if (resources.length === 0) {
+    throw new Error("User not found in organization");
+  }
+  return resources[0].id;
+}
+
+export async function getAdminRoleIdInOrganization(organizationId) {
+
+  const token = await getOrganizationToken(organizationId);
+  const response = await axios.get(
+    `${ASGARDEO_BASE_URL}/o/scim2/v2/Roles`,
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+        Accept: "application/json",
+      },
+      params: {
+          filter: `displayName eq Business Administrator`,
+        },
+      httpsAgent: agent,
+    }
+  );
+  const resources = response.data.Resources || [];
+  if (resources.length === 0) {
+    throw new Error("Admin role not found in organization");
+  }
+  return resources[0].id;
+}
+
+export async function addUserToAdminRole(organizationId, roleId, userId) {
+  
+  const token = await getOrganizationToken(organizationId);
+  const response = await axios.patch(
+    `${ASGARDEO_BASE_URL}/o/scim2/v2/Roles/${roleId}`,
+    {
+      Operations: [
+        {
+          op: "add",
+          path: "users",
+          value: [
+            {
+              value: userId,
+            },
+          ],
+        },
+      ],
+    },
+    {
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+      },
+      httpsAgent: agent, // Attach the custom agents
+    }
+  );
+  return response.data;
+}
